Corregí los atributos de la clase persona(Modelos), además agregué funciones al archivo Datos.py

+++ unidad 1/Clase6Victor/back/api/models/persona.py
@@ -3,8 +3,8 @@ from django.db import models
 class Persona(models.Model):
 	nombre 				= models.TextField() #'CharField' Campo de texto
 	apellido 				= models.TextField()
-	estatura 		= models.IntegerField()
-	peso 		= models.IntegerField()
+	estatura 		= models.DecimalField(max_digits=4, decimal_places=2)
+	peso 		= models.DecimalField(max_digits=4, decimal_places=2)
 	
 	def __str__(self):
 		return self.nombre
